+++ src/api-webhook.js
@@ -6,23 +6,21 @@ import log from "./log";
  * Handles web-hook callbacks from Github.
  *
  * @param {Object} settings:
- *                  - apps:       Collection of apps to start.
- *                  - manifest:   A manifest object.
+ *                  - mainApi:    The main API.
  *
  */
 export default (settings = {}) => {
-  const { apps, manifest } = settings;
+  const { mainApi } = settings;
   return {
     post(req, res) {
       // Extract data.
       const data = req.body;
       const branch = (data.ref && R.last(data.ref.split("/"))) || data.repository.default_branch;
       const repo = data.repository.full_name;
-      const modified = (data.head_commit && data.head_commit.modified) || [];
 
       // Match apps that reside within the repo that Github posted.
       const isRepoMatch = (app) => app.repo.name === repo && app.branch === branch;
-      const matchingApps = R.filter(isRepoMatch, apps);
+      const matchingApps = R.filter(isRepoMatch, mainApi.apps);
 
       // Update any matching apps.
       if (matchingApps.length > 0) {