+++ src/main-start.js
@@ -2,8 +2,7 @@ import Promise from "bluebird";
 import gateway from "./gateway";
 import log from "./log";
 import pm2 from "./pm2";
-import { promises, sortAppsByRoute } from "./util";
-import { DEFAULT_GATEWAY_PORT } from "./const";
+import { promises } from "./util";
 
 
 
@@ -12,7 +11,6 @@ import { DEFAULT_GATEWAY_PORT } from "./const";
  * Starts the gateway and all registered apps.
  *
  * @param {Object} settings:
- *                  - apps:           Collection of apps to start.
  *                  - manifest:       A manifest object.
  *                  - update:         Function that invokes the update method.
  *                  - port:           The port to start the gateway on.
@@ -22,7 +20,7 @@ import { DEFAULT_GATEWAY_PORT } from "./const";
  * @return {Promise}
  */
 export default (settings = {}) => {
-  let { apps, update, manifest, port, publishEvent, mainApi } = settings;
+  let { update, manifest, port, publishEvent, mainApi } = settings;
 
   return new Promise((resolve, reject) => {
     Promise.coroutine(function*() {
@@ -35,7 +33,7 @@ export default (settings = {}) => {
         yield manifest.update().catch(err => reject(err));
       }
 
-      if (apps.length === 0) {
+      if (mainApi.apps.length === 0) {
         log.warn("WARNING: No apps have been registered. Make sure a manifest has been set.");
 
       } else {
@@ -47,9 +45,8 @@ export default (settings = {}) => {
         }
 
         // Start the gateway and each app.
-        apps = sortAppsByRoute(apps);
-        yield gateway.start({ apps, port, manifest, publishEvent, mainApi }).catch(err => reject(err));
-        const { results: items } = yield promises(apps.map(app => app.start().catch(err => reject(err))));
+        yield gateway.start({ port, manifest, publishEvent, mainApi }).catch(err => reject(err));
+        const { results: items } = yield promises(mainApi.apps.map(app => app.start().catch(err => reject(err))));
 
         // Log status.
         log.info("");
@@ -57,8 +54,10 @@ export default (settings = {}) => {
         log.info(`Gateway running on port:${ port }`);
         log.info("");
         items.forEach(item => {
-            const version = item.version ? ` (v${ item.version })` : "";
-            log.info(` - '${ item.id }'${ version } routing '${ item.route }' => port:${ item.port }`);
+            if (item && item.exists) {
+              const version = item.version ? ` (v${ item.version })` : "";
+              log.info(` - '${ item.id }'${ version } routing '${ item.routes.toString() }' => port:${ item.port }`);
+            }
         });
         log.info("");