+++ src/route.js
@@ -19,9 +19,10 @@ const parse = (route) => {
 
   // URL path.
   let urlPath = R.takeLast(parts.length - 1, parts).join("/").trim();
-  urlPath = isEmpty(urlPath) ? undefined : urlPath;
-  if (urlPath) {
-    urlPath = urlPath.endsWith("/") ? urlPath : urlPath + "/";
+  urlPath = urlPath.replace(/\/$/, "");
+  urlPath = isEmpty(urlPath) ? "*" : urlPath;
+  if (urlPath !== "*") {
+    urlPath += "/";
   }
 
   return {
@@ -39,12 +40,13 @@ const parse = (route) => {
       path = formatMatchValue(path);
       if (path) { path = path.replace(/^\//, ""); }
       const isWildcardDomain = this.domain === "*";
+      const isWildcardPath = this.path === "*";
 
       if (!isWildcardDomain) {
         if (domain !== this.domain) { return false; }
       }
 
-      if (this.path) {
+      if (!isWildcardPath) {
         if (!path) { return false; }
         if (path && !(path + "/").startsWith(this.path)) { return false; }
       }
@@ -65,5 +67,17 @@ const parse = (route) => {
 
 
 
+/**
+ * Parses all routes within the given value(s).
+ * @param {String|Array} route: The Route or Routes to parse.
+ * @return {Array}.
+ */
+const parseAll = (route) => {
+  if (!R.is(Array, route)) { route = [route]; }
+  return R.map(parse, route);
+};
 
-export default { parse };
+
+
+
+export default { parse, parseAll };